src/utils/bybit/requirements.txt
requests
ujson
websocket-client
aiohttp